Add double-quotes to path names in cmd files.
When we call IF EXISTS and other commands in a batch file, and the given path has spaces then this command will silently fail. We enclose the passed path with double quotes so that the condition executes successfully. This commit will prevent build.cmd failing when %LocalAppData% path contains spaces (as mine did).